ref #164: Overwrite is now a three states field.
OVERWRITE, SKIP, FAIL where SKIP and FAIL behave the same way for single
output tasks. This way I only changed the type from boolean to enum and
it seemed little invasive given the field is buried in the inheritance
tree. Most of the changes are at client level, so the CLI has changed a
bit and sejda.com and PDFsam will also need to change.